feat(venue): remove VenueOffers old
